rename old struct
