+++ Gopkg.toml
@@ -30,10 +30,6 @@
   version = "0.7.0"
 
 [[constraint]]
-  name = "github.com/google/go-github"
-  version = "15.0.0"
-
-[[constraint]]
   name = "github.com/gorilla/mux"
   version = "1.6.2"
