+++ icode/infra/git/icode_git_store_test.go
@@ -16,10 +16,19 @@
 
 package git_test
 
-/*
+import (
+	"errors"
+	"os"
+	"testing"
+
+	"github.com/it-chain/engine/icode"
+	"github.com/it-chain/engine/icode/infra/git"
+	"github.com/stretchr/testify/assert"
+)
+
 func TestICodeGitStoreApi_Clone(t *testing.T) {
 	baseTempPath := "./.tmp"
-	sshPath := ""
+	sshPath := "./id_rsa"
 	os.RemoveAll(baseTempPath)
 	defer os.RemoveAll(baseTempPath)
 
@@ -30,8 +39,8 @@ func TestICodeGitStoreApi_Clone(t *testing.T) {
 		OutputErr   error
 	}{
 		"success": {
-			InputGitURL: "git@github.com:it-chain/tesseract.git",
-			OutputMeta:  &icode.Meta{RepositoryName: "tesseract", GitUrl: "git@github.com:it-chain/tesseract.git", Path: baseTempPath + "/" + "tesseract"},
+			InputGitURL: "git@github.com:hea9549/test_icode.git",
+			OutputMeta:  &icode.Meta{RepositoryName: "test_icode", GitUrl: "git@github.com:hea9549/test_icode.git", Path: baseTempPath + "/" + "test_icode"},
 			OutputErr:   nil,
 		},
 		"fail": {
@@ -41,8 +50,7 @@ func TestICodeGitStoreApi_Clone(t *testing.T) {
 		},
 	}
 
-	icodeApi, err := git.NewICodeGitStoreApi("testid", "testpw")
-	assert.NoError(t, err)
+	icodeApi := git.NewRepositoryService()
 
 	for testName, test := range tests {
 		t.Logf("Running %s test, case: %s", t.Name(), testName)
@@ -58,70 +66,5 @@ func TestICodeGitStoreApi_Clone(t *testing.T) {
 		//then
 		assert.Equal(t, test.OutputMeta, meta)
 		assert.Equal(t, test.OutputErr, err)
-
-	}
-
-}
-
-func TestNewICodeGitStoreApi(t *testing.T) {
-	type Input struct {
-		Id string
-		Pw string
-	}
-
-	//given
-	tests := map[string]struct {
-		Input  Input
-		Output string
-	}{
-		"validAccountCase": {
-			Input: Input{
-				// for test, write valid github id,pw
-				Id: "validId",
-				Pw: "validPw",
-			},
-			Output: "",
-		},
-		"invalidAccountCase": {
-			Input: Input{
-				Id: "invalidId",
-				Pw: "invalidPw",
-			},
-			Output: "GET https://api.github.com/user: 401 Bad credentials []",
-		},
-	}
-
-	for testName, test := range tests {
-		t.Logf("Running %s test, case: %s", t.Name(), testName)
-		//when
-		_, err := git.NewICodeGitStoreApi(test.Input.Id, test.Input.Pw)
-
-		if err != nil {
-			assert.Equal(t, err.Error(), test.Output)
-		} else {
-			assert.Equal(t, "", test.Output)
-		}
 	}
 }
-
-//todo push를 어떻게 확인할지. 단순 레포 리스트만확인하면 createRepo테스트임. push 함수의 err만 체크해도되는지?
-func TestICodeGitStoreApi_Push(t *testing.T) {
-	validId := "validId"
-	validPw := "validPw"
-	baseTempPath := "./.tmp"
-	sshPath := ""
-	os.RemoveAll(baseTempPath)
-	defer os.RemoveAll(baseTempPath)
-
-	storeApi, err := git.NewICodeGitStoreApi(validId, validPw)
-
-	assert.NoError(t, err)
-	meta, err := storeApi.Clone("1", baseTempPath, "git@github.com:it-chain/heimdall.git", sshPath)
-	err = storeApi.Push(*meta)
-	assert.NoError(t, err)
-
-	ctx := context.Background()
-	_, err = storeApi.AuthClient.Repositories.Delete(ctx, validId, meta.RepositoryName)
-	assert.NoError(t, err)
-}
-*/
